requirements.txt
boto3
pandas
numpy
mysql-connector-python
beautifulsoup4
requests
lxml